updated copies in hero-bootcamps and card bootcamp JS also, name bootcamp Python
