+++ config/bootcamps/hero-data.json
@@ -7,7 +7,7 @@
     "information": [
       {
         "title": "Inicio:",
-        "text": "08 de Febrero de 2021"
+        "text": "26 de Abril de 2021"
       },{
         "title":"Modalidad: Live",
         "text": "Clases remotas"
@@ -19,7 +19,7 @@
     ]
   },
   "python": {
-    "title": "Bootcamp Web developer + Data",
+    "title": "Bootcamp Back End con Python",
     "subtitle": "Javascript + Python",
     "description": "Transforma tu vida programando. En 21 semanas tendrás las habilidades para programar una aplicación web desde cero.",
     "information": [
